moved PayPalTransaction to namespace Transaction #42
